fix(build): remove SITE_DIR constant
Remove `SITE_DIR` constant from cli/build.go. The constant and it's use
in the code assumes that the program is run against a `site/`
folder in the current directory. In cases where the program binary is
not in the same directory or the path supplied with the `build` option
in the cli does not have the name `site` this will cause the program to
fail. To fix this pass the `buildPath` supplied as the value for the
`build` option to the `renderFiles` function.